1/15 풀 받기전에 푸쉬
