+++ src/main/java/project/web/mvc/controller/AdminUserController.java
@@ -2,8 +2,8 @@ package project.web.mvc.controller;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -13,7 +13,6 @@ import lombok.RequiredArgsConstructor;
 import project.web.mvc.domain.Userdb;
 import project.web.mvc.service.AdminUserService;
 import project.web.mvc.service.AuthorityService;
-import project.web.mvc.service.OnLectureService;
 import project.web.mvc.service.UserdbService;
 
 
@@ -24,9 +23,7 @@ public class AdminUserController {
 	
 	private final AdminUserService adminuserService;
 	private final UserdbService userdbService;
-	private final OnLectureService onLectureService;
 	private final AuthorityService authorityService;
-	
 	
 	
 	@RequestMapping("")
@@ -70,4 +67,35 @@ public class AdminUserController {
 		System.out.println("key : " + key + "//////" + "value :" + value);
 		adminuserService.selectByKey(key, value);
 	}
+	
+	
+//	회원가입 처리
+	@RequestMapping("/signUp")
+	@Transactional
+	@ResponseBody
+	public int execSignUp(String userdbEmail, String userdbNickname, String userdbPassword) {
+
+		Userdb userdb = new Userdb();
+		userdb.setUserdbEmail(userdbEmail);
+		userdb.setUserdbNickname(userdbNickname);
+		userdb.setUserdbPassword(userdbPassword);
+		//0L = admin, 1L = member
+		userdb.setAuthority(0L);
+		//회원가입
+		userdbService.insert(userdb);
+		//회원가입 후 authority db 저장
+		authorityService.insert(userdb);
+		System.out.println("회원가입끝");
+		//1 : 성공 , 0: 실패
+		int result = 1;
+		return result;
+	}
+	
+	   //signupForm으로 가기
+	   @RequestMapping("/signUpForm")
+	   public String adminSignUpForm() {
+		   return"admin/user/signUpForm";
+	   }
+
+   
 }
