--- src/main/java/project/web/mvc/controller/MemberController.java
@@ -1,8 +1,0 @@
-package project.web.mvc.controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class MemberController {
-
-}
